* commit 'dac80f5b0dbf840e15d6d2eb8eed2e7ee7113b66':
  Fix -Wconversion warnings in framework/common.

+++ framework/common/tcuTextureUtil.cpp
@@ -901,14 +901,14 @@ void scale (const PixelBufferAccess& dst, const ConstPixelBufferAccess& src, Sam
 	{
 		for (int y = 0; y < dst.getHeight(); y++)
 		for (int x = 0; x < dst.getWidth(); x++)
-			dst.setPixel(src.sample2D(sampler, filter, (x+0.5f)*sX, (y+0.5f)*sY, 0), x, y);
+			dst.setPixel(src.sample2D(sampler, filter, ((float)x+0.5f)*sX, ((float)y+0.5f)*sY, 0), x, y);
 	}
 	else
 	{
 		for (int z = 0; z < dst.getDepth(); z++)
 		for (int y = 0; y < dst.getHeight(); y++)
 		for (int x = 0; x < dst.getWidth(); x++)
-			dst.setPixel(src.sample3D(sampler, filter, (x+0.5f)*sX, (y+0.5f)*sY, (z+0.5f)*sZ), x, y, z);
+			dst.setPixel(src.sample3D(sampler, filter, ((float)x+0.5f)*sX, ((float)y+0.5f)*sY, ((float)z+0.5f)*sZ), x, y, z);
 	}
 }
